property-crawler-worker/requirements.txt
beautifulsoup4
boto3
meilisearch
Pillow
pydantic==1.10.11
pymongo
celery[redis]
Celery
psycopg2-binary
flower
random_user_agent
redis
pytz
prisma
pillow-avif-plugin
